% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cost_of_equity.R
\name{cost_of_equity}
\alias{cost_of_equity}
\title{Download the Cost of Equity}
\usage{
cost_of_equity()
}
\value{
A \code{tibble}
}
\description{
Get the cost of equity for different sectors according to the CAPM methodology.
}
\examples{
#cost_of_equity()
}
