% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dividend_yield.R
\name{dividend_yield}
\alias{dividend_yield}
\title{Download the Dividend Yield}
\usage{
dividend_yield()
}
\value{
A \code{tibble}
}
\description{
Get the ratio of the dividends over the stock prices in a given period.
}
\examples{
#dividend_yield()
}
