% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iliquidity_index.R
\name{iliquidity_index}
\alias{iliquidity_index}
\title{Download the Illiquidity Index}
\usage{
iliquidity_index()
}
\value{
A \code{tibble}
}
\description{
Get a response on how the market moves in response to its traded volume.
}
\examples{
#iliquidity_index()
}
