% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nefin_data.R
\name{nefin_data}
\alias{nefin_data}
\title{nefin_data}
\usage{
nefin_data(data, category = "unique", portfolio_weight = "Equally",
  frequency = "")
}
\arguments{
\item{data}{a string containing the type of data to download. Available options: Cost of Equity ("COE"), Dividend Yield ("dividend_yield"),
Illiquidity Index ("illiquidity_index"), Loan Fees ("loan_fees"), Portfolios ("portfolios"), Risk Factors ("risk_factors"),
Short Interest ("short_interest"), Volatility Index ("volatility_index").}

\item{category}{a string (or a vector of string) containing the category of data to download. Only necessary for Cost of Equity,
Porfolios, Risk Factors and Volatility Index data type. See 'Details'.}

\item{portfolio_weight}{a string. Choose "Equally" (default option) for Equally Weighted Returns and "Value" for Value Weighted Returns}

\item{frequency}{an optional string. Choose "month" for monthly data and "year" for annual data.
Only available for Cost of Equity, Risk Factors, Dividend Yield, Illiquidity Index, Loan Fees and Short Interest. See 'Details'.}
}
\value{
A DataFrame containing the selected data
}
\description{
nefin_data returns a DataFrame with financial data from NEFIN website.
}
\details{
If you choose Cost of Equity, Portfolios, Risk Factors or Volatility Index in category,
you need to choice one or more available options for category. If you use the default option,
the function will download all available options.
Use a string for one option and a vector of string for more than one option. \cr
\emph{Cost of Equity}: Basic Products ("Products"), Construction ("Construction"), Consumer ("Consumer"), Energy ("Energy"),
Finance("Finance"), Manufacturing ("Manufacturin"), Other ("Other").
\cr See details of each category \href{http://www.nefin.com.br/cost_of_capital.html}{here} \cr

\cr \emph{Portfolios}: 3 portfolios sorted by size ("size"), 3 portfolios sorted by book-to-market("btm"),
3 portfolios sorted by momentum ("momentum"), 3 portfolios sorted by illiquidity ("illiquidity"),
4 portfolios sorted by size and by book-to-market ("size_btm"), 4 portfolios sorted by size and by momentum ("size_momentum"),
4 portfolios sorted by size and illiquidity ("size_illiquidity")
7 portfolios sorted by industry ("industry").
\cr See more details \href{http://www.nefin.com.br/portfolios.html}{here} \cr

\cr \emph{Risk Factors}: Market Factor ("Rm_minus_Rf"), Small Minus Big ("SMB"), High Minus Low ("HML"),
Winners Minus Losers ("WML"), Illiquid Minus Liquid ("IML"), Risk Free ("Risk_free").
\cr See details \href{http://www.nefin.com.br/risk_factors.html}{here}. \cr

\cr \emph{Volatility Index}: IVol-Br Volatility Index ("IVolBR"), Variance Premium ("variance_premium"), Risk Aversion ("risk_aversion")

\cr \cr If you choose Portfolios in category, you also need to choose an option for 'portfolio_weigth".
The default option is "Equally" for Equally Weighted Returns and the other option is "Value" for Value Weighted Returns.

\cr \cr For frequency, the default option are:
\cr Cost of Equity: monthly
\cr Dividend Yield: weekly
\cr Illiquidity Index: monthly
\cr Loan Fees: monthly
\cr Portfolios: daily
\cr Risk Factors: daily
\cr Short Interest: weekly
\cr Volatility Index: daily for IVol-Br and weekly for Variance Premium
\cr \cr For Cost of Equity, Dividend Yield, Illiquidity Index, Loan Fees, Portfolios and Short Interest,
you can choose "year" for annual data or "month" for monthly data.

\cr \cr \cr For information about the methodology used, click \href{http://www.nefin.com.br/Metodologia/Methodology.pdf}{here}.
}
\examples{
dataframe <- nefin_data(data = "portfolios", category = "size", portfolio_weigth = "Value", frequency = "year")
dataframe <- nefin_data(data = "COE", category = "Products")
dataframe <- nefin_data(data = "illiquidity_index")
dataframe <- nefin_data(data = "loan_fees")
dataframe <- nefin_data(data = "dividend_yield")
dataframe <- nefin_data(data = "short_interest")
dataframe <- nefin_data(data = "risk_factors", category = "Rm_minus_Rf")
}
