% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolios.R
\name{portfolios_sorted_by_size}
\alias{portfolios_sorted_by_size}
\alias{portfolios_sorted_by_book_to_market}
\alias{portfolios_sorted_by_momentum}
\alias{portfolios_sorted_by_illiquidity}
\alias{portfolios_sorted_by_industry}
\title{Download Factor Portfolios}
\usage{
portfolios_sorted_by_size()

portfolios_sorted_by_book_to_market()

portfolios_sorted_by_momentum()

portfolios_sorted_by_illiquidity()

portfolios_sorted_by_industry()
}
\value{
A \code{tibble}
}
\description{
Portfolios of Different Sorts
}
\examples{
#portfolios_sorted_by_size()
}
