% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/short_interest.R
\name{short_interest}
\alias{short_interest}
\title{Download the Short-Interest}
\usage{
short_interest()
}
\value{
A \code{tibble}
}
\description{
Get the ratio between the number of shorted stock and the stock outstanding
for the Market.
}
\examples{
#short_interest()
}
